+++ changelog.md
@@ -1,3 +1,7 @@
+# Upcoming
+* Fixed daggers not usable as furnace fuel
+* Fixed daggers not breaking some blocks faster like sword
+
 # 5.12.1.0
 * Added missing sockets (Rune enchanting) to Shortbows
 * Added chains repair kits